Rework focus of model interpretation section
We currently don't have a good formula to calculate the
distribution of the null hypothesis when comparing a homogeneous
vs a heterogeneous model with the same fixed effects (but different
variance components). To solve this, focus on exploring the
difference in variance components by study qualitatively, without a
formal statistical test. Remove the homoskedastic null model from
the exercise and only fit a heterogeneous one, since we don't have
a good way to compare the two models. Because the exercise no longer
compares the two models, reorder the sections to start by looking at
the fixed effects, and then look at the variance components, which
is a more logical order.