Save the phenotype annotation for use in later exercises
Rework the order of the prep work so that the saved annotation
does not have sample identifiers, because they need to use a
sample-subject mapping to do genotype-phentoype association tests.
That will be covered in a different session; it makes more sense to
have the phenotype data stored without a sample identifier.