Use new analysis_pipeline repo location
The analysis_pipeline repo provides the source for a lot of the
test data used in the phenotype harmonization markdown doc. It was
moved from Stephanie's github account to the UW-GAC organization on
github. This meant that the repo paths needed to be updated to
point to the correct location. Additionally, it also caused the
download.file command to fail because github was using a redirect
to point to the file. To fix that, I had to add the extra = "-L"
keyword argument to download.file.