Fix url for downloading grm from analysis_pipeline repo
Use only the base filename of the grm file instead of the entire
grmfile variable, because the "data" dir doesn't exist in the
analysis_pipeline repo. Remove the extra = "-L" since it is not
needed on amazon (only in my cloned repo which is using packrat for
packages... who knows).